changed totalPrice from a int to a double as getPrice from product is a double

+++ src/main/java/wms/PurchaseOrder.java
@@ -41,7 +41,7 @@ public class PurchaseOrder {
         this.status = newStatus;
     }
 
-    public int totalPrice() {
+    public double totalPrice() {
         return quantity * product.getPrice();
     }
 }
